Removed unnecessary process termination
Removed the usage of NtTerminateProcess in the Utilities.c file. This change was made as the process termination was redundant and unnecessary, making the code cleaner and more efficient. The process will now only be terminated if the hSection is NULL, streamlining the termination procedure and simplifying the logic within the code block.