Refactor README for improved clarity and organization
This commit rearranges some content in the README file to better communicate the project's goal and setup instructions. Notably, a note about the broken support for 3rd party agents in the Havoc 0.6 version is moved closer to the 'Setup' heading, making it more visible to readers. A reference to the corresponding Havoc PR#379 was also added for readers seeking more information.